1.3.4 Move objectify types to type file
